Update FogCommand methods. Change int values to float.

+++ src/main/java/creman/fog/api/FogUtil.java
@@ -4,19 +4,17 @@ import creman.fog.ClientProxy;
 
 public class FogUtil
 {
-    public static void SetFogColor (int red, int green, int blue)
+    public static void SetFogColor(float red, float green, float blue)
     {
         ClientProxy.red = red;
         ClientProxy.green = green;
         ClientProxy.blue = blue;
     }
-
-    public static void SetFogDensity (float density)
+    public static void SetFogDensity(float density)
     {
         ClientProxy.density = density;
     }
-
-    public static void SetFog (int red, int green, int blue, float density)
+    public static void SetFog(float red, float green, float blue, float density)
     {
         SetFogColor(red, green, blue);
         SetFogDensity(density);
